requirements.txt
autopep8==1.5.3
boto3==1.13.18
botocore==1.16.18
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-cors-headers==3.3.0
django-silk==4.0.1
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils==0.15.2
drf-writable-nested==0.6.0
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
MarkupSafe==1.1.1
pilkit==2.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
drf-yasg==1.17.1
